// src/app/interfaces/evidence/evidence-view.interface.ts
export interface IEvidenceView {
  id?: number;
  name?: string,
  publicId?: string,
  eTag?: string,
  url?: string,
  securedUrl?: string,
  mimeType?: string;
  type?: string;
  Program?: string;
  ProgramId?: number;
  AssessmentId?: number;
  MyClassId?: number;
  InstructorId?: number;
  ProgramCourseId?: number;
  Course?: string;
  ProgramSopiId?: number;
  Sopi?: string;
}